+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## v1.0.3.4 ()
 - Select multiple rows in backup list for deletion
+- Detect Windows 7 and disable update check
 
 ## v1.0.3.2 (30.5.2020)
 - Add icons to all menu items